Revert "Merged Post and Search Page"
This reverts commit 563a20083460a5f05b0a3d4206121204489bbb60.
