+++ lib/screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:campus_ola/decorations/home_screen_style.dart';
 import 'package:campus_ola/models/post_model.dart';
-import 'package:campus_ola/screens/create_post_and_search/create_post.dart';
-import 'package:campus_ola/screens/create_post_and_search/search_page.dart';
+import 'package:campus_ola/screens/create_post/create_post.dart';
+import 'package:campus_ola/screens/search_page/search_page.dart';
 import 'package:campus_ola/widgets/home/post_widget.dart';
 import 'package:flutter/material.dart';
